More .d.ts
