+++ src/game/typechart.d.ts
@@ -1,0 +1,3 @@
+export class Typechart {
+  compare(move: string, target: Array<any> | string): number; // TODO
+}
