+++ src/interfaces/cli.d.ts
@@ -1,0 +1,5 @@
+export class Interactive {
+  challenge(): void;
+  onLobbyEnter(): void;
+  onUpdateUser(nick: any, status: any): any; // TODO
+}
