+++ src/log.d.ts
@@ -1,0 +1,3 @@
+export class Log {
+  setLogLevel(lvl: number): void;
+}
