+++ src/model/side.d.ts
@@ -1,0 +1,7 @@
+export class Side {
+  constructor();
+
+  data(): any; // TODO
+  digest(action: string): void;
+  remove(action: string): void;
+}
