+++ src/report.d.ts
@@ -1,0 +1,6 @@
+export class Report {
+  constructor();
+
+  data(): Array<object>; // TODO
+  win(victor: string, store: BattleStore, matchid: string): Report;
+}
