+++ src/timer.d.ts
@@ -1,0 +1,6 @@
+export class Timer {
+  constructor();
+
+  after(cb: any, seconds: number): void; // TODO
+  ping(): void;
+}
